--- src/component/atoms/button/SecondaryButton.jsx
@@ -1,14 +1,0 @@
-import styled from "styled-components";
-
-import { BaseButton } from "./BaseButton";
-
-export const SecondaryButton = (props) => {
-  const { children, onClick } = props;
-  return <SButton onClick={onClick}>{children}</SButton>;
-};
-
-const SButton = styled(BaseButton)`
-  background: #fff100;
-  color: #333;
-  margin-right: 20px;
-`;
